// src/components/HtmlFileView/HtmlDocument/types.ts
type ColorScheme = {
  background: string;
  bridge: string;
  tagBlanket: string;
  tagName: string;
  attributeName: string;
  attributeValue: string;
  textNode: string;
  cssTagName: string;
  cssAttributeName: string;
  cssAttributeValue: string;
  cssBlanket: string;
};

export type { ColorScheme };